Igo8 warning cleanup.

+++ trunk/gpsbabel/igo8.c
@@ -103,19 +103,19 @@ static void igo8_check_type_sizes()
 	if (sizeof(igo8_point) != 12)
 	{
 		fatal(MYNAME ": igo8_point is %ld bytes instead of the required 12.\n",
-			sizeof(igo8_point));
+			(long) sizeof(igo8_point));
 	}
 
 	if (sizeof(igo8_information_block) != 12)
 	{
 		fatal(MYNAME ": igo8_information_block is %ld bytes instead of the required 12.\n",
-			sizeof(igo8_information_block));
+			(long) sizeof(igo8_information_block));
 	}
 
 	if (sizeof(igo8_id_block) != 20)
 	{
 		fatal(MYNAME ": igo8_id_block is %ld bytes instead of the required 20.\n",
-			sizeof(igo8_id_block));
+			(long) sizeof(igo8_id_block));
 	}
 }
